% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{p_show}
\alias{p_show}
\title{Convert a raw p value to conventional notation}
\usage{
p_show(x, max_digits = 3, signif = 2, nonbreaking = TRUE)
}
\arguments{
\item{x}{A \emph{p} value.}

\item{max_digits}{The maximum number of decimal places to show.}

\item{signif}{The number of significant digits to include.}

\item{nonbreaking}{Should spaces be non-breaking?}
}
\value{
A string in the form of \emph{p} = x or \emph{p} < x
}
\description{
This function converts a raw \emph{p} value with any number of decimal places to
a string that is used conventionally to report \emph{p} values in the form of
\emph{p} = x or \emph{p} < x.
}
\examples{
p_show(0.0125623123, max_digits = 2)

test = t.test(c(1,2,3), c(5,4,3))
p_show(test$p.value)
}
