% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{power_val}
\alias{power_val}
\title{Function to report high values as powers of 10}
\usage{
power_val(x)
}
\arguments{
\item{x}{A large number.}
}
\value{
A string in the form of a $\\cdot$ 10^b^
}
\description{
This function converts a large raw number to a string of a multiplicative
expression with 10^x^.
}
\examples{
power_val(exp(20))
}
