% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{report}
\alias{report}
\title{Function to return specific cells of an ANOVA table by term}
\usage{
report(results, term, which_value, digits = 3, print_name = TRUE)
}
\arguments{
\item{results}{An ANOVA table or lm object or data frame.}

\item{term}{The term of which to report the value.}

\item{which_value}{The type of value to report (\emph{p} value, t statistic, ...). Either "p" or the name of the statistic ("F", "t", ...). Currently only F supported}

\item{digits}{The number of digits to which to round the value.}

\item{print_name}{Should the name of the statistic ("F", "t", ...) be printed? \strong{NOT IMPLEMENTED YET}}
}
\value{
A string specifying degrees of freedom and test statistic.
}
\description{
Conveniently report single values (like \emph{p} value or t statistic) of an ANOVA
table by term. \strong{UNFINISHED!!!}
}
