% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{to_named_list}
\alias{to_named_list}
\title{Function to convert a data frame to a list named by a column.}
\usage{
to_named_list(df, by)
}
\arguments{
\item{df}{A data frame.}

\item{by}{Name of identifying column}
}
\value{
The value of the specified cell.
}
\description{
Convenient method to access cells within a data frame by an identifying
column value.
}
\examples{
df = data.frame(ID = c("Item1", "Item2"), First_Value = c(2, 3), Second_Value = c("A", "B"))
results = to_named_list(df, by = "ID")

results$Item1$First_Value
}
